+++ folio-service-tools-spring-dev/src/test/java/org/folio/spring/tools/kafka/KafkaAdminServiceTest.java
@@ -71,6 +71,20 @@ class KafkaAdminServiceTest {
   }
 
   @Test
+  void createKafkaTopics_positive_withTopicPrefix() {
+    when(kafkaProperties.getTopicPrefix()).thenReturn("myprefix");
+    when(kafkaProperties.getTopics()).thenReturn(List.of(
+      FolioKafkaProperties.KafkaTopic.of("topic4", 5, null)));
+
+    kafkaAdminService.createTopics(TEST_TENANT);
+    verify(kafkaAdmin).initialize();
+
+    var beansOfType = applicationContext.getBeansOfType(NewTopic.class);
+    assertThat(beansOfType).containsValue(
+      new NewTopic("folio.test_tenant.myprefix.topic4", Optional.of(5), Optional.empty()));
+  }
+
+  @Test
   void deleteKafkaTopics_positive() {
     FolioKafkaProperties.KafkaTopic kafkaTopic = new FolioKafkaProperties.KafkaTopic();
     kafkaTopic.setName("test_topic");
@@ -91,6 +105,30 @@ class KafkaAdminServiceTest {
 
     verify(kafkaClient).listTopics();
     verify(kafkaClient).deleteTopics(List.of("folio.test_tenant.test_topic"));
+  }
+
+  @Test
+  void deleteKafkaTopics_positive_withTopicPrefix() {
+    when(kafkaProperties.getTopicPrefix()).thenReturn("myprefix");
+    FolioKafkaProperties.KafkaTopic kafkaTopic = new FolioKafkaProperties.KafkaTopic();
+    kafkaTopic.setName("test_topic");
+
+    var future = KafkaFuture.completedFuture(Set.of("folio.test_tenant.myprefix.test_topic"));
+    var listTopicResult = mock(ListTopicsResult.class);
+    when(listTopicResult.names()).thenReturn(future);
+
+    when(kafkaProperties.getTopics()).thenReturn(List.of(kafkaTopic));
+    var kafkaClient = mock(AdminClient.class);
+    try (var ignored = mockStatic(AdminClient.class, invocation -> kafkaClient)) {
+      var deleteTopicsResult = mock(DeleteTopicsResult.class);
+      when(kafkaClient.listTopics()).thenReturn(listTopicResult);
+      when(kafkaClient.deleteTopics(anyCollection())).thenReturn(deleteTopicsResult);
+      when(deleteTopicsResult.all()).thenReturn(KafkaFuture.completedFuture(mock(Void.class)));
+      kafkaAdminService.deleteTopics(TEST_TENANT);
+    }
+
+    verify(kafkaClient).listTopics();
+    verify(kafkaClient).deleteTopics(List.of("folio.test_tenant.myprefix.test_topic"));
   }
 
   @Test